Refactored voting form code into a separate component and fixed styles issues
