+++ src/components/App/App.tsx
@@ -1,7 +1,15 @@
 import { Route, Routes, useNavigate } from 'react-router-dom';
 import { useEffect } from 'react';
+import {
+  ProtectedRouteElementForAuthorized,
+  ProtectedRouteElementForUnauthorized,
+} from './ProtectedRoute';
+
 import authStore from '@/stores/AuthStore';
+import alertStore from '@/stores/AlertStore';
+
 import { getUserMe } from '@/api/authService';
+
 import ContentBlock from '../ContentBlock/ContentBlock';
 import DepartmentPanel from '../../Pages/DepartmentPanel/DepartmentPanel';
 import DocumentTake from '@/Pages/DocumentTake/DocumentTake';
@@ -10,18 +18,15 @@ import Footer from '../Footer/Footer';
 import Header from '../Header/Header';
 import Alert from '../Alert/Alert';
 import { observer } from 'mobx-react-lite';
-import alertStore from '@/stores/AlertStore';
-import styles from './app.module.css';
 import { Paths } from '@/enums/Paths';
 import Auth from '../Auth/Auth';
-import {
-  ProtectedRouteElementForAuthorized,
-  ProtectedRouteElementForUnauthorized,
-} from './ProtectedRoute';
 import FormLogin from '../Auth/Forms/FormLogin';
 import FormRegistration from '../Auth/Forms/FormRegistration';
+
 import DocumentPanel from '@/Pages/DocumentPanel/DocumentPanel';
 import NotFoundPage from '../../Pages/NotFoundPage/NotFoundPage';
+
+import styles from './app.module.css';
 
 const App: React.FC = () => {
   const navigate = useNavigate();
@@ -67,10 +72,6 @@ const App: React.FC = () => {
             </>
           ) : (
             <>
-              <Route
-                path={Paths.DOCUMENTS_VOTE}
-                element={'Компонент голосования за документ (Юзер)'}
-              />
               <Route path={Paths.DOCUMENTS_TAKE} element={<DocumentTake />} />
               <Route
                 path={`${Paths.DOCUMENTS_TAKE}/:id/:appId?/:appItemId?`}
