+++ src/components/ContentBlock/DocumentModal/documentModal.module.css
@@ -91,80 +91,3 @@
 .dot.active {
   background-color: rgb(145, 145, 145);
 }
-
-.dataList__img {
-  width: 1.5rem;
-  height: 1.5rem;
-}
-
-.dataList__controls {
-  margin-top: 2rem;
-  padding: 4px;
-  font-size: 1.2rem;
-  background-color: rgb(224, 224, 224);
-  display: inline-flex;
-  align-items: center;
-  gap: 0px;
-  margin-bottom: 15px;
-  cursor: pointer;
-  transition: all 0.2s ease;
-}
-
-.dataList__controls:hover {
-  opacity: 0.75;
-}
-
-.dataList__button {
-  border: none;
-  background-color: transparent;
-  cursor: pointer;
-}
-
-.vote {
-  text-align: center;
-}
-
-.vote__button {
-  font-size: 1rem;
-  background-color: rgb(224, 224, 224);
-  border: none;
-  margin-top: 2rem;
-  padding: 7px 10px 7px 10px;
-  transition: all 0.2s ease;
-}
-
-.vote__button:hover {
-  opacity: 0.75;
-}
-
-.title {
-  margin-bottom: 10px;
-}
-
-.voteButtons {
-  display: flex;
-  justify-content: center;
-  gap: 10px;
-}
-
-.voteButton {
-  padding: 10px;
-  cursor: pointer;
-  border: none;
-  color: white;
-  font-weight: bold;
-  transition: background-color 0.3s;
-}
-
-.voteButton.choice {
-  background-color: rgb(224, 224, 224); 
-}
-
-
-.voteButton.active {
-  background-color: #800080; /* Фиолетовый цвет для активной кнопки */
-}
-
-.voteButton:disabled {
-  background-color: white; 
-}
